+++ CHANGELOG.md
@@ -21,6 +21,8 @@
 
 - `gallery`
   - [#3165](https://github.com/wix/ricos/pull/3165) fix height issues on scroll direction change
+- `file-upload`
+  - [#3167](https://github.com/wix/ricos/pull/3167) fix icon colors wiring
 
 ### :house: Internal